scraper-service/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
ical>=1.0.0
pymongo>=4.6.0
motor>=3.3.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
aiofiles>=23.2.0

# Development & Code Quality
ruff>=0.1.0
flake8>=6.1.0
black>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
